update events-handlers to v3
