--- events-producer/README.md
@@ -1,7 +1,0 @@
-## Build
-
-Use the following command to build the **events-producer** image:
-
-```shell
-docker build -t <tag> .
-```
