--- eventsource/example/deploy/eventsource-handler.yaml
@@ -1,29 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: eventsource-handler
-  labels:
-    app: eventsourcehandler
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: eventsourcehandler
-  template:
-    metadata:
-      labels:
-        app: eventsourcehandler
-      annotations:
-        dapr.io/enabled: "true"
-        dapr.io/app-id: "eventsource-handler"
-        dapr.io/app-port: "5050"
-        dapr.io/log-as-json: "true"
-        dapr.io/app-protocol: "grpc"
-    spec:
-      containers:
-        - name: eventsource
-          image: openfunctiondev/eventsource-handler:v1
-          imagePullPolicy: IfNotPresent
-          env:
-            - name: CONFIG
-              value: "eyJldmVudFNvdXJjZUNvbXBvbmVudCI6ImthZmthLWV2ZW50c291cmNlIiwiZXZlbnRCdXNDb21wb25lbnQiOiJuYXRzLWV2ZW50YnVzIiwiZXZlbnRCdXNUb3BpYyI6InRvcGljMSIsInNpbmtDb21wb25lbnQiOiJodHRwLXNpbmsifQ=="
