--- eventsource/example/deploy/sink-component.yaml
@@ -1,11 +1,0 @@
-apiVersion: dapr.io/v1alpha1
-kind: Component
-metadata:
-  name: http-sink
-spec:
-  type: bindings.http
-  version: v1
-  metadata:
-    - name: url
-      # This is the access address of a Knative Service
-      value: http://function-sample-serving-qrdx8-ksvc-fwml8.default.192.168.0.5.sslip.io
