--- eventsource/function/go.mod
@@ -1,8 +1,0 @@
-module openfunction.io/eventsource
-
-go 1.15
-
-require (
-	github.com/OpenFunction/functions-framework-go v0.0.0-20210922063920-81a7b2951b8a
-	github.com/sirupsen/logrus v1.8.1
-)
