+++ eventsource/function/v2/go.mod
@@ -1,0 +1,8 @@
+module openfunction.io/eventsource
+
+go 1.15
+
+require (
+	github.com/OpenFunction/functions-framework-go v0.1.1
+	github.com/sirupsen/logrus v1.8.1
+)
