+++ eventsource/function/v3/go.mod
@@ -1,0 +1,8 @@
+module openfunction.io/eventsource
+
+go 1.16
+
+require (
+	github.com/OpenFunction/functions-framework-go v0.2.0
+	k8s.io/klog/v2 v2.30.0
+)
