--- eventsource/handler/go.mod
@@ -1,5 +1,0 @@
-module eventsourcehandlers.io/bindings
-
-go 1.15
-
-require github.com/dapr/go-sdk v1.1.0
