+++ examples/events-producer/go.mod
@@ -1,0 +1,7 @@
+module openfunction.io/events-producer
+
+go 1.16
+
+require (
+	github.com/OpenFunction/functions-framework-go v0.2.0
+)
