+++ examples/producer.yaml
@@ -1,0 +1,52 @@
+apiVersion: core.openfunction.io/v1beta1
+kind: Function
+metadata:
+  name: producer
+spec:
+  version: "v1.0.0"
+  port: 8080
+  image: openfunctiondev/events-producer:latest
+#  imageCredentials:
+#    name: push-secret
+#  build:
+#    builder: openfunction/builder-go:latest
+#    env:
+#      FUNC_NAME: "Producer"
+#      FUNC_CLEAR_SOURCE: "true"
+#    srcRepo:
+#      url: "https://github.com/OpenFunction/events-handlers.git"
+#      sourceSubPath: "examples/events-producer"
+  serving:
+    template:
+      containers:
+        - name: function
+          imagePullPolicy: Always
+    runtime: "async"
+    inputs:
+      - name: cron
+        component: cron
+    outputs:
+      - name: target
+        component: kafka-server
+        operation: "create"
+    bindings:
+      cron:
+        type: bindings.cron
+        version: v1
+        metadata:
+          - name: schedule
+            value: "@every 2s"
+      kafka-server:
+        type: bindings.kafka
+        version: v1
+        metadata:
+          - name: brokers
+            value: "kafka-server-kafka-brokers:9092"
+          - name: topics
+            value: "demo"
+          - name: consumerGroup
+            value: "event-producer"
+          - name: publishTopic
+            value: "demo"
+          - name: authRequired
+            value: "false"
