--- trigger/example/config/go.mod
@@ -1,3 +1,0 @@
-module eventsourcehandlers.io/bindings
-
-go 1.15
