--- trigger/example/deploy/events-producer.yaml
@@ -1,30 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: events-producer
-  labels:
-    app: eventsproducer
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: eventsproducer
-  template:
-    metadata:
-      labels:
-        app: eventsproducer
-      annotations:
-        dapr.io/enabled: "true"
-        dapr.io/app-id: "events-producer"
-        dapr.io/log-as-json: "true"
-        dapr.io/app-protocol: "grpc"
-    spec:
-      containers:
-        - name: producer
-          image: openfunctiondev/events-producer:latest
-          imagePullPolicy: IfNotPresent
-          env:
-            - name: TARGET_NAME
-              value: "nats-eventbus"
-            - name: TARGET_TOPIC
-              value: "default-my-eventsource-sample-one"
