--- trigger/example/deploy/trigger-component.yaml
@@ -1,17 +1,0 @@
-apiVersion: dapr.io/v1alpha1
-kind: Component
-metadata:
-  name: trigger
-spec:
-  type: pubsub.natsstreaming
-  version: v1
-  metadata:
-    - name: natsURL
-      value: "nats://nats.default:4222"
-    - name: natsStreamingClusterID
-      value: "stan"
-      # below are subscription configuration.
-    - name: subscriptionType
-      value: queue
-    - name: consumerID
-      value: trigger
