--- trigger/example/deploy/trigger-handler.yaml
@@ -1,29 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: trigger
-  labels:
-    app: trigger
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: trigger
-  template:
-    metadata:
-      labels:
-        app: trigger
-      annotations:
-        dapr.io/enabled: "true"
-        dapr.io/app-id: "trigger"
-        dapr.io/log-as-json: "true"
-        dapr.io/app-protocol: "grpc"
-        dapr.io/app-port: "5050"
-    spec:
-      containers:
-        - name: trigger
-          image: openfunctiondev/trigger-handler:v1
-          imagePullPolicy: IfNotPresent
-          env:
-            - name: CONFIG
-              value: "eyJidXNDb21wb25lbnQiOiJ0cmlnZ2VyIiwiYnVzVG9waWMiOlt7Im5hbWUiOiJBIiwibmFtZXNwYWNlIjoiZGVmYXVsdCIsImV2ZW50U291cmNlIjoibXktZXZlbnRzb3VyY2UiLCJldmVudCI6InNhbXBsZS1vbmUifSx7Im5hbWUiOiJCIiwibmFtZXNwYWNlIjoiZGVmYXVsdCIsImV2ZW50U291cmNlIjoibXktZXZlbnRzb3VyY2UiLCJldmVudCI6InNhbXBsZS10d28ifV0sInN1YnNjcmliZXJzIjp7IkEgXHUwMDI2XHUwMDI2IEIiOnsidG9waWMiOiJtZXRyaWNzIn0sIkIiOnsic2lua0NvbXBvbmVudCI6Imh0dHAtc2luayJ9fSwicG9ydCI6IjUwNTAifQ=="
