--- trigger/function/go.mod
@@ -1,13 +1,0 @@
-module openfunction.io/trigger
-
-go 1.15
-
-require (
-	github.com/OpenFunction/functions-framework-go v0.0.0-20210922063920-81a7b2951b8a
-	github.com/dapr/go-sdk v1.2.0
-	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
-	github.com/google/cel-go v0.7.3
-	github.com/sirupsen/logrus v1.8.1
-	google.golang.org/genproto v0.0.0-20201204160425-06b3db808446
-	google.golang.org/protobuf v1.25.0
-)
