+++ trigger/function/v2/go.mod
@@ -1,0 +1,13 @@
+module openfunction.io/trigger
+
+go 1.15
+
+require (
+	github.com/OpenFunction/functions-framework-go v0.1.1
+	github.com/dapr/go-sdk v1.2.0
+	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
+	github.com/google/cel-go v0.7.3
+	github.com/sirupsen/logrus v1.8.1
+	google.golang.org/genproto v0.0.0-20201204160425-06b3db808446
+	google.golang.org/protobuf v1.25.0
+)
