+++ trigger/function/v3/go.mod
@@ -1,0 +1,13 @@
+module openfunction.io/trigger
+
+go 1.16
+
+require (
+	github.com/OpenFunction/functions-framework-go v0.2.0
+	github.com/cloudevents/sdk-go/v2 v2.4.1
+	github.com/dapr/go-sdk v1.3.1
+	github.com/google/cel-go v0.9.0
+	google.golang.org/genproto v0.0.0-20210831024726-fe130286e0e2
+	google.golang.org/protobuf v1.27.1
+	k8s.io/klog/v2 v2.30.0
+)
