--- trigger/handler/.dockerignore
@@ -1,2 +1,0 @@
-# comment
-Dockerfile
