--- trigger/handler/go.mod
@@ -1,9 +1,0 @@
-module eventsourcehandlers.io/trigger
-
-go 1.15
-
-require (
-	github.com/dapr/go-sdk v1.2.0
-	github.com/google/cel-go v0.7.3
-	google.golang.org/genproto v0.0.0-20201204160425-06b3db808446
-)
